+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.9.0] - 2022-07-18
+
 ### Added
 
 - `SearchableList`: Allow showing a copy symbol for the identifier (enabled for collections and processes)
@@ -536,7 +538,8 @@ First stable release.
 Please see the [Releases](https://github.com/Open-EO/openeo-vue-components/releases) for changelogs prior to v1.0.0.
 
 
-[Unreleased]: https://github.com/Open-EO/openeo-vue-components/compare/v2.8.1...HEAD
+[Unreleased]: https://github.com/Open-EO/openeo-vue-components/compare/v2.9.0...HEAD
+[2.9.0]: https://github.com/Open-EO/openeo-vue-components/compare/v2.8.1...v2.9.0
 [2.8.1]: https://github.com/Open-EO/openeo-vue-components/compare/v2.8.0...v2.8.1
 [2.8.0]: https://github.com/Open-EO/openeo-vue-components/compare/v2.7.1...v2.8.0
 [2.7.1]: https://github.com/Open-EO/openeo-vue-components/compare/v2.7.0...v2.7.1